Refactors 'index.js' to several modules.
Moves expense transformer to 'expense_transformer.js' module.
Moves file reading to 'read_expense_files.js' module.
Moves API calls to 'botler_api.js' module.
